refact and doc
